#113 introduced floating point types
