+++ lisa/lisa-imp/src/main/java/it/unive/lisa/imp/IMPCodeMemberVisitor.java
@@ -78,7 +78,6 @@ import it.unive.lisa.imp.expressions.IMPStringLiteral;
 import it.unive.lisa.imp.expressions.IMPSub;
 import it.unive.lisa.imp.expressions.IMPTrueLiteral;
 import it.unive.lisa.imp.types.ClassType;
-import it.unive.lisa.imp.types.FloatType;
 import it.unive.lisa.program.Global;
 import it.unive.lisa.program.SourceCodeLocation;
 import it.unive.lisa.program.annotations.Annotations;
@@ -108,6 +107,7 @@ import it.unive.lisa.program.cfg.statement.VariableRef;
 import it.unive.lisa.type.Type;
 import it.unive.lisa.type.Untyped;
 import it.unive.lisa.type.common.BoolType;
+import it.unive.lisa.type.common.Float32;
 import it.unive.lisa.type.common.Int32;
 import it.unive.lisa.util.datastructures.graph.AdjacencyMatrix;
 
@@ -630,7 +630,7 @@ class IMPCodeMemberVisitor extends IMPParserBaseVisitor<Object> {
 		else if (ctx.INT() != null)
 			return Int32.INSTANCE;
 		else
-			return FloatType.INSTANCE;
+			return Float32.INSTANCE;
 	}
 
 	@Override
