+++ lisa/lisa-imp/src/main/java/it/unive/lisa/imp/expressions/IMPFloatLiteral.java
@@ -1,13 +1,13 @@
 package it.unive.lisa.imp.expressions;
 
-import it.unive.lisa.imp.types.FloatType;
 import it.unive.lisa.program.SourceCodeLocation;
 import it.unive.lisa.program.cfg.CFG;
 import it.unive.lisa.program.cfg.statement.Literal;
+import it.unive.lisa.type.common.Float32;
 
 /**
  * An IMP {@link Literal} representing a constant floating point value.
- * Instances of this literal have a {@link FloatType} static type.
+ * Instances of this literal have a {@link Float32} static type.
  * 
  * @author <a href="mailto:luca.negrini@unive.it">Luca Negrini</a>
  */
@@ -23,6 +23,6 @@ public class IMPFloatLiteral extends Literal {
 	 * @param value      the constant value represented by this literal
 	 */
 	public IMPFloatLiteral(CFG cfg, String sourceFile, int line, int col, float value) {
-		super(cfg, new SourceCodeLocation(sourceFile, line, col), value, FloatType.INSTANCE);
+		super(cfg, new SourceCodeLocation(sourceFile, line, col), value, Float32.INSTANCE);
 	}
 }
